CustomLabelInput now sits in its own file
