+++ keepsake-app/db.json
@@ -24,9 +24,15 @@
       "noteID": 1,
       "userID": 1,
       "title": "Groceries",
-      "content": ["apples", "eggs", "pasta"],
+      "content": [
+        "apples",
+        "eggs",
+        "pasta"
+      ],
       "category": "shopping list",
-      "labels": [1],
+      "labels": [
+        1
+      ],
       "isChecklist": true,
       "id": "255d"
     },
@@ -36,7 +42,10 @@
       "title": "Meeting Notes",
       "content": "Discuss marketing strategy and finalise budget allocations.",
       "category": "work",
-      "labels": [2, 3],
+      "labels": [
+        2,
+        3
+      ],
       "isChecklist": false,
       "id": "5ff1"
     },
@@ -44,9 +53,15 @@
       "noteID": 3,
       "userID": 2,
       "title": "Home Repairs",
-      "content": ["Fix leaky tap", "Paint living room", "Replace broken tile"],
+      "content": [
+        "Fix leaky tap",
+        "Paint living room",
+        "Replace broken tile"
+      ],
       "category": "household",
-      "labels": [4],
+      "labels": [
+        4
+      ],
       "isChecklist": true,
       "id": "6c7f"
     },
@@ -56,7 +71,10 @@
       "title": "Book Recommendations",
       "content": "Check out 'Atomic Habits' and 'Deep Work' for productivity tips.",
       "category": "personal growth",
-      "labels": [5, 6],
+      "labels": [
+        5,
+        6
+      ],
       "isChecklist": false,
       "id": "3d7c"
     },
@@ -64,9 +82,15 @@
       "noteID": 5,
       "userID": 3,
       "title": "Weekend Plans",
-      "content": ["Hiking trip", "Try new sushi place", "Movie night"],
+      "content": [
+        "Hiking trip",
+        "Try new sushi place",
+        "Movie night"
+      ],
       "category": "leisure",
-      "labels": [7],
+      "labels": [
+        7
+      ],
       "isChecklist": true,
       "id": "2305"
     },
@@ -76,7 +100,9 @@
       "title": "Project Ideas",
       "content": "Brainstorm app ideas for personal finance tracking.",
       "category": "work",
-      "labels": [8],
+      "labels": [
+        8
+      ],
       "isChecklist": false,
       "id": "5a81"
     }
@@ -129,6 +155,18 @@
       "userID": 3,
       "labelName": "Finance",
       "id": "9adb"
+    },
+    {
+      "id": "9c65",
+      "labelName": "Fun"
+    },
+    {
+      "id": "d3dd",
+      "labelName": "Fun"
+    },
+    {
+      "id": "7287",
+      "labelName": "Groceries"
     }
   ]
 }
