+++ keepsake-app/src/components/Sidebar.tsx
@@ -1,7 +1,8 @@
 import React from "react";
 import "./Sidebar.css";
-import { useState, useEffect } from "react";
+import { useState } from "react";
 import { label } from "../App.tsx";
+import { CustomLabelInput } from "./customLabelInput.tsx";
 
 interface sidebarProps {
   labels: label[];
@@ -15,45 +16,6 @@ export default function Sidebar({ labels, setLabels }: sidebarProps) {
     setIsEditLabels(true);
   }
 
-  const Input = () => {
-    const [inputValue, setInputValue] = useState("");
-
-    const handleKeyDown = async (
-      event: React.KeyboardEvent<HTMLInputElement>
-    ) => {
-      if (event.key === "Enter" && inputValue.trim() !== "") {
-        try {
-          const newLabel = { labelName: inputValue };
-          const res = await fetch(`http://localhost:3000/labels`, {
-            method: "POST",
-            headers: { "Content-Type": "application/json" },
-            body: JSON.stringify(newLabel),
-          });
-
-          if (!res.ok) throw new Error("Failed to post label");
-          const data = await res.json();
-          setLabels((prev: label[]) => [...prev, data]); //our post request only returns the new label not the list of labels
-          setInputValue("");
-          setIsEditLabels(false);
-        } catch (err) {
-          console.error("Error posting label:", err);
-        }
-      }
-    };
-
-    return (
-      <label>
-        Custom label:
-        <input
-          type="text"
-          value={inputValue}
-          onChange={(e) => setInputValue(e.target.value)}
-          onKeyDown={handleKeyDown}
-        />
-      </label>
-    );
-  };
-
   return (
     <div className="sidebar">
       <button className="edit-labels-btn" onClick={handleEditLabelsClick}>
@@ -61,7 +23,10 @@ export default function Sidebar({ labels, setLabels }: sidebarProps) {
       </button>
       {isEditLabels && (
         <>
-          <Input />
+          <CustomLabelInput
+            setIsEditLabels={setIsEditLabels}
+            setLabels={setLabels}
+          />
           <select>
             {labels.map((label, idx) => (
               <option key={idx} value={label.labelName}>
